Cleanup Sector Map View
Integrate the hyperjump planner into the sectormap module
Convert sectormap window sections to tabs
